# cmd/api/docs/swagger.yaml
definitions:
...
  request.Coupon:
    properties:
      discountpercent:
        type: number
      expirationdate:
        type: string
      maximumdiscountprice:
        type: number
      minimumpurchaseamount:
        type: number
      usagelimit:
        type: integer
    required:
    - discountpercent
    - expirationdate
    - maximumdiscountprice
    - minimumpurchaseamount
    - usagelimit
    type: object
...
  respondse.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/coupon/AddCoupons:
    post:
      description: Admin can add coupon
      operationId: AddCoupon
      parameters:
      - description: Input true info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.Coupon'
      responses:
        "200":
          description: Successfully productItem added to cart
        "400":
          description: can't add the product item into cart
      security:
      - ApiKeyAuth: []
      summary: api for add Coupons for ecommerce
      tags:
      - Coupon
  /admin/coupon/Delete/{CouponID}:
    delete:
      consumes:
      - application/json
      description: Admin can delete a coupon
      operationId: delete-coupon
      parameters:
      - description: CouponID
        in: path
        name: CouponID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can delete a coupon
      tags:
      - Coupon
  /admin/coupon/Update/{CouponID}:
    patch:
      consumes:
      - application/json
      description: Admin can update existing coupon
      operationId: update-coupon
      parameters:
      - description: CouponID
        in: path
        name: CouponID
        required: true
        type: integer
      - description: details of coupon to be updated
        in: body
        name: coupon_details
        required: true
        schema:
          $ref: '#/definitions/request.Coupon'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can update existing coupon
      tags:
      - Coupon
  /admin/coupon/Viewcoupon/{id}:
    get:
      consumes:
      - application/json
      description: Admins can see Coupons with coupon_id
      operationId: find-Coupon-by-id
      parameters:
      - description: CouponID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admins  can see Coupons with coupon_id
      tags:
      - Coupon
  /admin/coupon/couponlist:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all coupons
      operationId: List-all-coupons
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: for geting all order status list
      tags:
      - Coupon
...
  /coupon/apply/{code}:
    patch:
      consumes:
      - application/json
      description: User can apply coupon to the cart
      operationId: applay-coupon-to-cart
      parameters:
      - description: code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: User can apply a coupon to the cart
      tags:
      - UsersCart
  /coupon/coupons:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all coupons in user side
      operationId: List-all-coupons-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: for geting all coupon list
      tags:
      - Users
...
swagger: "2.0"